mcp/bk/pyproject.toml
[project]
name = "bk-mcp"
version = "0.1.0"
description = "MCP server exposing Buildkite CI helpers as typed tools."
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.scripts]
bk-mcp = "bk_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bk_mcp"]
